backend/requirements.txt
Flask==3.0.0
pandas==2.1.2
joblib==1.3.2
pydantic==2.4.2
mysql-connector-python==8.2.0 
gunicorn==0.24.0
python-dotenv==1.0.0
scikit-learn==1.2.2